Add files via upload
Web scraping of webshop WEBSHOP "www.möbelfreude.de" to display the links for each product with their own parameters. The project is done using BeautifulSoup to scrape the webpage and requests to download or make a connection to the webpage. Pandas is used to make a data frame and save the data frame into CSV. Tqdm library is used to check progress.